test(tracking): cover the new persisted display prefs, and move their test
The persistence anti-drift fixture didn't know about `showLine`/`showPoints`, so
`persist()` emitted them as undefined and `JSON.stringify` dropped them: both the
exact-blob assertion and the save→rehydrate round-trip passed without ever
touching them, and removing either from the schema or the persist callback would
have lost the user's choice with a green suite. They are non-default in the
fixture on purpose — the defaults would round-trip either way.

The reducer test also moves next to the reducer it exercises.
